REFACTOR: removing method of the dbservice and implementing in netflixService
